docs: Update LE/NGINX CSP to fix loading of external gravatar images
This adds `https://secure.gravatar.com` back to the CSP so that gravatar
images will load in Chromium-based browsers as well as Firefox.

It seems this was removed in favor of `https://gravatar.com` in PR #661
to fix issue #660. I am unsure why `https://gravatar.com` is also
needed, but have left it in the CSP until we have more information.