perf(discovery): improve discovery first page performance above the fold
This change aims to improve discovery page performance by
only hydrating above the fold content. The general use case
for machines that load this page tend not to show anything below
"trending". This change leans on intersection observer to
lazy hydrate the page. This means that as the user scrolls through
the discovery page, the content below the fold (the unrendered part
of the app) will fire the SWR requests and fetch data.

This change is aimed to free up network congestion that stalls the page
render. A secondary improvement, not in this diff, is to play with
the Recently Requested component as that may also be slowing down page
load.