fix: fixes RT ratings for Tvshows
The releaseYear that RTAlgolia returns in the contentResults is of type=number. This messes up the
TVshow RTratings as even though the TVShow was found in the contentResults when the year is
compared, it returns `undefined` thus throwing 404 RT ratings not found. This fix fixes this issue
by treating the releaseYear as a number type.

fix #3491